chore(imports): break circular dependency edges
Remove the circular import edges found in the SDK modifier nodes, mapgen compiler helpers, and mapgen op authoring types by moving shared types to lower-level owner modules and correcting a wrong SDK fill payload type.

This keeps behavior unchanged while preserving a one-way dependency graph across these surfaces: nodes may instantiate child nodes, compiler normalization consumes compiler error types from a lower helper, and op contract/envelope/strategy code share types through op/types instead of importing each other.

Verification:
- bun run --cwd packages/mapgen-core check
- bun run --cwd packages/sdk check
- bun test packages/mapgen-core/test/compiler/recipe-compile.test.ts packages/mapgen-core/test/authoring/authoring.test.ts
- bun run --cwd packages/sdk test
- targeted local circular-import scan for SDK nodes, mapgen compiler, and mapgen op authoring surfaces
- git diff --check
