+++ packages/mapgen-core/src/authoring/op/index.ts
@@ -3,9 +3,9 @@ export { createOp } from "./create.js";
 export { createStrategy } from "./strategy.js";
 export { opRef } from "./ref.js";
 
-export type { OpContract, StrategyConfigSchemas } from "./contract.js";
+export type { OpContract } from "./contract.js";
 export type { OpContractLike, OpStrategyId, OpTypeBag, OpTypeBagOf } from "./types.js";
-export type { DomainOp, DomainOpKind } from "./types.js";
+export type { DomainOp, DomainOpKind, StrategyConfigSchemas } from "./types.js";
 export type { OpRef } from "./ref.js";
 export type {
   OpStrategy,
