+++ packages/mapgen-core/src/compiler/normalize.ts
@@ -5,7 +5,7 @@ import type { DomainOpCompileAny, OpsById } from "../authoring/bindings.js";
 import { bindCompileOps, OpBindingError } from "../authoring/bindings.js";
 import { applySchemaDefaults } from "../authoring/schema.js";
 import type { StepOpsDecl } from "../authoring/step/ops.js";
-import type { CompileErrorItem } from "./recipe-compile.js";
+import type { CompileErrorItem } from "./errors.js";
 
 export type NormalizeCtx<TEnv = unknown, TKnobs = unknown> = Readonly<{ env: TEnv; knobs: TKnobs }>;
